requirements.txt
nltk==3.8.1
pipwin==0.5.0; platform_system == "Windows"
pytest==7.4.0
google-generativeai
python-dotenv
flask
